refactor scalar test for readability
